django-data-handler/requirements.txt
asn1crypto==0.24.0
boto3==1.7.4
botocore==1.10.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
Django==1.11
djangorestframework==3.6.2
docutils==0.14
future==0.16.0
gevent==1.2.2
greenlet==0.4.13
idna==2.6
jmespath==0.9.3
pprint==0.1
predix==1.0.0
psycopg2-binary==2.7.4
pycparser==2.18
python-dateutil==2.6.0
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
websocket==0.2.1
websocket-client==0.47.0
